+++ robotidy/api.py
@@ -37,7 +37,7 @@ class RobotidyAPI(Robotidy):
             src=(),
             exclude=exclude,
             extend_exclude=extend_exclude,
-            ignore_gitignore=False,
+            skip_gitignore=False,
             overwrite=False,
             show_diff=False,
             formatting_config=formatting_config,